Implement Timeout for AuthenticateAsClientAsync
See Issue #18. Implement a timeout wrapper around the call to the SslStream.AuthenticateAsClientAsync() because it doesn't take in a cancellation token itself. If the client attempts to connect to a server that does not implement TLS, a call to this method may wait forever because the server won't send back any data for the TLS handshake that this method is expecting. We'll use 100 seconds, which is the same as the default timeout for a WebRequest under a similar condition.