SyslogLoggerConfigurationExtensions Cleanup (#26)
See Issue #25 for more details. When using the `UdpSyslog()` extension method, if the `format` parameter was set to `SyslogFormat.RFC5424`, it would cause an `ArgumentNullException` later on when trying to send an event. The `messageIdPropertyName` optional parameter was defaulted to `null` in that case and would cause the exception when it was used to look up the property key in the `Dictionary<>`.

We now default the `messageIdPropertyName` parameter to "SourceContext" in all places necessary.

Refactored the `UdpSyslogReceiver` class to model it after the `TcpSyslogReceiver` class such that is will itself obtain a random port to listen on, instead of creating a separate `Socket` just to get a port. As well as support listening on both IPv4 and IPv6 via `DualMode`.

Also addressed Issue #20 where some Unit Tests had timeout values of 4,000 and 6,000 seconds. While they never had to actually wait that long, the code was misleading. Changed to use a computed value based upon the sending time of the `PeriodicBatchingSinkOptions`.