Graceful Fallback to IPv4 Only if IPv6 is Unavailable
In some environments, like AWS Lambda, IPv6 is not available. This is not the same as testing with IPv6 disabled, for example, which is all we've been able to do up to this point. If IPv6 is completely unavailable, then trying to create a socket with the `AddressFamily.InterNetworkV6` would throw an exception. We'll try and handle that gracefully with a check here by using the built-in `Socket.OSSupportsIPv6` property instead of relying on a `try/catch` block with an exception.

We'll assume that if only IPv4 is available, then the passed in host will resolve to an IPv4 address. No additional checks will be performed.