Issue #40: Invalid log levels
A new parameter to allow for a callback method that can be used to provide alternative mappings from Serilog `EventLogLevel` to Syslog `Severity`.

Offer a public alternate severity mapping that can be used as the new parameter value instead of having to write your own method.

Expose `SyslogFormatterBase.CalculatePriority()` as `virtual` so someone can easily override it and have complete control.