Correct Use of SslProtocols Enum
To address feedback given in #47, this is a breaking change for anyone using the `SyslogTcpConfig` or `SyslogLoggerConfigurationExtensions.TcpSyslog()` extension method. The `SecureProtocols` property/parameter has been replace with a simple boolean, `UseTls`. If you were previously setting the `SecureProtocols` property/parameter to a value such as `SslProtocols.Tls12`, then you would set the new `UseTls` property/parameter to `true`. If you were using `SslProtocols.None`, then you will pass in `false` for the new boolean value.

Also updated in the PR are the certificate based unit tests to cleanup/delete the private keys after they are used. Otherwise, they become orphaned on the hard drive. (That is for Windows specifically, not sure what happens on Linux.)

Lastly, the main project, Serilog.Sinks.Syslog, is set to target just the .NET Standard 2.0 framework. Because we do not have any framework specific code, this one build will allow it to be run under any .NET, on any platform, that supports .NET Standard 2.0. This include .NET Framework 4.6.2+, .NET Core2.0+, and .NET 5+, on Windows, Linux, or MacOS.