Able to reproduce the SslProtocols error for .NET Framework 4.6.2 on Windows Server 2019 using a test application. With the test application, we can set the `AppContext` switch programmatically at any time prior to the SSL connection and have the connection be successful.
Specifying the `AppContext` switch in the app.config file for these unit tests probably doesn't work like it did for the test application because these unit tests don't produce an executable, they just produce a DLL, for which app.config files don't apply.

Therefore, the only other thing I can think of is perhaps the conditional compilation that was being used around the `AppContext` to set the switch for only .NET 4.6.2 wasn't working. So we will try it without the conditional compilation and have that code always run. Since it's the default for all other .NET versions, it won't hurt.

Also, change back to specifying and building the main project for multiple targets. While a .NET Standard 2.0 assembly can be referenced by any of the supported .NET versions, doing so would cause other .NET Standard 2.0 references and assemblies to be included in the person's output, which is undesirable.