Support ClientCertificateSelectionCallback Issue #81
Add support for the TCP mutual authentication to allow the client authentication certificate to be dynamically selected, via a callback method, upon each SSL connection. This allows for the rotation of a certificate without having to recreate and initialize an entire new TcpSyslogConfig.

See https://github.com/IonxSolutions/serilog-sinks-syslog/issues/81.

The behavior of the CertificateProvider has essentially been updated to allow for the same thing. That is, if you specify a CertificateProvider, the Syslog Sink will now retrieve the Certificate property each time it establishes a connection. Before, it was only taking a copy of the certificate once, in the constructor, for the lifetime of the configuration.