Multi-line Exception Example and Unit Test Assembly Binding Workaround (#83)
* Multi-line Exception Example

A user asked if it was possible to log "the whole exception callstack to syslog". Here are a couple of unit tests showing that.

See https://github.com/IonxSolutions/serilog-sinks-syslog/issues/70.

* Unit Test Assembly Binding Workaround

Maybe it doesn't happen on every machine, but since Serilog.Sinks.PeriodicBatching was upgraded to 5.0.0, it references Serilog 4.0.0, which references System.Threading.Channels 8.0.0, which references System.Threading.Tasks.Extensions 4.5.4, which finally references System.Runtime.CompilerServices.Unsafe 6.0.0. But something else is referencing System.Runtime.CompilerServices.Unsafe 4.0.4.1. That causes an assembly load failure because there is no assembly binding redirects.

Using the AssemblyResolve event in code to act as the binding redirect solves the problem as there are no compatibility issues between the versions for our purposes.